pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "latticing"
dynamic = ["version"]
description = "Using Behavior Latticing to infer user motivation from unstructured interactions"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.96.0",
    "build>=1.4.4",
    "dotenv>=0.9.9",
    "google-genai>=1.47.0",
    "openai>=2.32.0",
    "plotly>=5.0.0",
    "pyyaml>=6.0.3",
    "tenacity>=9.1.2",
    "together>=2.9.0",
    "tqdm>=4.0.0",
    "twine>=6.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[tool.hatch.version]
path = "src/latticing/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/latticing"]

[tool.hatch.build.targets.sdist]
exclude = [
    "/website",
    "/node_modules",
    "/.venv",
    "/.pytest_cache",
    "/Agent_Aimer*",
    "/docs",
    "/examples",
    "/tests"
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
src = ["src"]
line-length = 88
